Completed a "first stage" of the demonstration of using the Fortran wrapper.
The purpose of this code is to demonstrate how computational elements can be
created when using the underlying Fluent case file reader library API from a
Fortran code. With the present data structures actual solvers can be built!

Added a file that is presently under development (part of a different library
distribution) and is used to build computational elements for use in finite
element and finite volume computations. That library's API is used to create
the elements that are represented in the Fluent case file that is read by the
present library API through its Fortran wrapper.

The driver code (in this step) has been hard-wired to output tetrahedral
elements in a Tecplot file format. This means that meaningful output will be
generated, if at all, only when the Fluent case file contains only
tetrahedral elements. Error messages and premature termination will occur if
other elements are present in the file.
IN 2017/10/16