Modified the test3.f driver code to accommodate changes in the mesh library code. In hte process, added some deallocations for arrays on termination that were missing.
Started to create the GLSQ method of gradient calculation.

In the process, a problem with the elements construction was found, and the
volumes appear to be negative. Further diagnostics reveal that there is a
problem with how elements are assembled, and manifested in CELLS.dat. There
needs to be more investigation to correct the problem!
IN 2017/11/29